+++ lsm/memtable.hpp
@@ -7,20 +7,20 @@ using json = nlohmann:: json;
 
 class Memtable {
 public:
-    explicit Memtable(int maxKvPairs) : maxKvPairs(maxKvPairs) {}
+    explicit Memtable(size_t maxKvPairs) : maxKvPairs(maxKvPairs) {}
     ~Memtable() {};
 
     bool put(KEY_t key, VAL_t value);
     std::unique_ptr<VAL_t> get(KEY_t key) const;
     std::map<KEY_t, VAL_t> range(KEY_t start, KEY_t end) const;
     void clear();
-    int size() const;
+    size_t size() const;
     std::map<KEY_t, VAL_t> getMap() const;    
     long getMaxKvPairs() const;
     json serialize() const;
     void deserialize(const json& j);
 private:
-    long maxKvPairs;
+    size_t maxKvPairs;
     std::map<KEY_t, VAL_t> table_;  
 };
